Update the translations.
Add the Spanish translation in place of the missing Latin American Spanish translation.
The Hebrew locale was originally marked as HE, but now we use IW. Copy the IW directory to the HE directory to make sure that it won't stop working for users who installed it under HE locale.
Add the Slovak translation, which was missing.