Remove the check for large notification index.
In shouldShowNotification_, we immediately return true for large indices. This is the correct behavior, since it allows us to try again if the extension uninstallation failed.

However, in maybeShowNotification_, we set the index to zero if it is out of range, which means that we reset the timeline. This CL removes the upper bound check in maybeShowNotification_, so that this doesn't happen.

To test that the whole process works, we add a mock of Chrome extension APIs.