Copy the timestamp from the synced storage to the local one as well.
To differentiate between a fresh installation (empty local storage) and a state when this extension was woken up after missing a synced uninstallation request (non-empty local storage). In the latter case, we uninstall the extension immediately.

Note that no timestamp is actually needed; in the local storage, we only actually need a boolean flag. However, the timestamp is added for consistency, for compliance with the agreed specification, and for possible future purposes (if this fix was not enough to cope with the syncing bug).